+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # SDRSharp.SatnogsTracker Changelog
+## 1.5.2
+* Resample Audio Stream to 48Khz using Naudio
+* Will stream for AF recordings.
+* Switch to celestrak amateur.txt for now
 ## 1.5.1
 * Fixed starting point of the red bars
 * Added settings option to update TLEs and re-download Sats and Transmitters